pyproject.toml
[tool.poetry]
name = "subsidies-dot-io"
version = "0.1.0"
description = ""
authors = ["dsaks9 <delonsaks@berkeley.edu>"]
readme = "README.md"
packages = [{include = "subsidies_dot_io"}, {include = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.2.3"
numpy = "^2.1.3"
jupyter = "^1.1.1"
firecrawl-py = "^1.6.1"
pydantic = "^2.10.2"
asyncio = "^3.4.3"
python-dotenv = "^1.0.1"
llama-index-core = "0.12.0"
llama-index-vector-stores-qdrant = "^0.4.0"
llama-index-postprocessor-cohere-rerank = "^0.3.0"
llama-index-embeddings-cohere = "^0.4.0"
llama-index = "0.12.0"
typing-extensions = "^4.12.2"
fastembed = "^0.4.2"
llama-index-program-openai = "^0.3.1"
arize-phoenix-otel = "^0.6.1"
openinference-instrumentation-llama-index = "^3.0.4"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
python-docx = "^1.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
